+++ js/src/frontend/BytecodeEmitter.cpp
@@ -31,9 +31,11 @@
 #include "ds/Nestable.h"
 #include "frontend/BytecodeControlStructures.h"
 #include "frontend/EmitterScope.h"
+#include "frontend/ForOfLoopControl.h"
 #include "frontend/IfEmitter.h"
 #include "frontend/Parser.h"
 #include "frontend/TDZCheckCache.h"
+#include "frontend/TryEmitter.h"
 #include "frontend/TokenStream.h"
 #include "vm/Debugger.h"
 #include "vm/GeneratorObject.h"
@@ -67,463 +69,6 @@ ParseNodeRequiresSpecialLineNumberNotes(ParseNode* pn)
 {
     return pn->getKind() == ParseNodeKind::While || pn->getKind() == ParseNodeKind::For;
 }
-
-// Class for emitting bytecode for blocks like try-catch-finally.
-//
-// Usage: (check for the return value is omitted for simplicity)
-//
-//   `try { try_block } catch (ex) { catch_block }`
-//     TryEmitter tryCatch(this, TryEmitter::Kind::TryCatch,
-//                         TryEmitter::ControlKind::Syntactic);
-//     tryCatch.emitTry();
-//     emit(try_block);
-//     tryCatch.emitCatch();
-//     emit(ex and catch_block); // use JSOP_EXCEPTION to get exception
-//     tryCatch.emitEnd();
-//
-//   `try { try_block } finally { finally_block }`
-//     TryEmitter tryCatch(this, TryEmitter::Kind::TryFinally,
-//                         TryEmitter::ControlKind::Syntactic);
-//     tryCatch.emitTry();
-//     emit(try_block);
-//     // finally_pos: The "{" character's position in the source code text.
-//     tryCatch.emitFinally(Some(finally_pos));
-//     emit(finally_block);
-//     tryCatch.emitEnd();
-//
-//   `try { try_block } catch (ex) {catch_block} finally { finally_block }`
-//     TryEmitter tryCatch(this, TryEmitter::Kind::TryCatchFinally,
-//                         TryEmitter::ControlKind::Syntactic);
-//     tryCatch.emitTry();
-//     emit(try_block);
-//     tryCatch.emitCatch();
-//     emit(ex and catch_block);
-//     tryCatch.emitFinally(Some(finally_pos));
-//     emit(finally_block);
-//     tryCatch.emitEnd();
-//
-class MOZ_STACK_CLASS TryEmitter
-{
-  public:
-    enum class Kind {
-        TryCatch,
-        TryCatchFinally,
-        TryFinally
-    };
-
-    // Syntactic try-catch-finally and internally used non-syntactic
-    // try-catch-finally behave differently for 2 points.
-    //
-    // The first one is whether TryFinallyControl is used or not.
-    // See the comment for `controlInfo_`.
-    //
-    // The second one is whether the catch and finally blocks handle the frame's
-    // return value.  For syntactic try-catch-finally, the bytecode marked with
-    // "*" are emitted to clear return value with `undefined` before the catch
-    // block and the finally block, and also to save/restore the return value
-    // before/after the finally block.
-    //
-    //     JSOP_TRY
-    //
-    //     try_body...
-    //
-    //     JSOP_GOSUB finally
-    //     JSOP_JUMPTARGET
-    //     JSOP_GOTO end:
-    //
-    //   catch:
-    //     JSOP_JUMPTARGET
-    //   * JSOP_UNDEFINED
-    //   * JSOP_SETRVAL
-    //
-    //     catch_body...
-    //
-    //     JSOP_GOSUB finally
-    //     JSOP_JUMPTARGET
-    //     JSOP_GOTO end
-    //
-    //   finally:
-    //     JSOP_JUMPTARGET
-    //   * JSOP_GETRVAL
-    //   * JSOP_UNDEFINED
-    //   * JSOP_SETRVAL
-    //
-    //     finally_body...
-    //
-    //   * JSOP_SETRVAL
-    //     JSOP_NOP
-    //
-    //   end:
-    //     JSOP_JUMPTARGET
-    //
-    // For syntactic try-catch-finally, Syntactic should be used.
-    // For non-syntactic try-catch-finally, NonSyntactic should be used.
-    enum class ControlKind {
-        Syntactic,
-        NonSyntactic
-    };
-
-  private:
-    BytecodeEmitter* bce_;
-    Kind kind_;
-    ControlKind controlKind_;
-
-    // Track jumps-over-catches and gosubs-to-finally for later fixup.
-    //
-    // When a finally block is active, non-local jumps (including
-    // jumps-over-catches) result in a GOSUB being written into the bytecode
-    // stream and fixed-up later.
-    //
-    // For non-syntactic try-catch-finally, all that handling is skipped.
-    // The non-syntactic try-catch-finally must:
-    //   * have only one catch block
-    //   * have no catch guard
-    //   * have JSOP_GOTO at the end of catch-block
-    //   * have no non-local-jump
-    //   * don't use finally block for normal completion of try-block and
-    //     catch-block
-    //
-    // Additionally, a finally block may be emitted for non-syntactic
-    // try-catch-finally, even if the kind is TryCatch, because GOSUBs are not
-    // emitted.
-    Maybe<TryFinallyControl> controlInfo_;
-
-    // The stack depth before emitting JSOP_TRY.
-    int depth_;
-
-    // The source note index for SRC_TRY.
-    unsigned noteIndex_;
-
-    // The offset after JSOP_TRY.
-    ptrdiff_t tryStart_;
-
-    // JSOP_JUMPTARGET after the entire try-catch-finally block.
-    JumpList catchAndFinallyJump_;
-
-    // The offset of JSOP_GOTO at the end of the try block.
-    JumpTarget tryEnd_;
-
-    // The offset of JSOP_JUMPTARGET at the beginning of the finally block.
-    JumpTarget finallyStart_;
-
-#ifdef DEBUG
-    // The state of this emitter.
-    //
-    // +-------+ emitTry +-----+   emitCatch +-------+      emitEnd  +-----+
-    // | Start |-------->| Try |-+-------+-->| Catch |-+->+--------->| End |
-    // +-------+         +-----+ |       ^   +-------+ |  ^          +-----+
-    //                           |       |             |  |
-    //                           |  +----+-------------+  +----+
-    //                           |  |                          |
-    //                           |  v emitFinally +---------+  |
-    //                           +->+------------>| Finally |--+
-    //                                            +---------+
-    enum class State {
-        // The initial state.
-        Start,
-
-        // After calling emitTry.
-        Try,
-
-        // After calling emitCatch.
-        Catch,
-
-        // After calling emitFinally.
-        Finally,
-
-        // After calling emitEnd.
-        End
-    };
-    State state_;
-#endif
-
-    // Waterfox: We haven't yet removed conditional catch clauses, so we still
-    // need some sort of miniature state tracking even outside of debug builds.
-    bool emittedTry_;
-
-    bool hasCatch() const {
-        return kind_ == Kind::TryCatch || kind_ == Kind::TryCatchFinally;
-    }
-    bool hasFinally() const {
-        return kind_ == Kind::TryCatchFinally || kind_ == Kind::TryFinally;
-    }
-
-  public:
-    TryEmitter(BytecodeEmitter* bce, Kind kind, ControlKind controlKind)
-      : bce_(bce),
-        kind_(kind),
-        controlKind_(controlKind),
-        depth_(0),
-        noteIndex_(0),
-        tryStart_(0),
-        emittedTry_(false)
-#ifdef DEBUG
-      , state_(State::Start)
-#endif
-    {
-        if (controlKind_ == ControlKind::Syntactic)
-            controlInfo_.emplace(bce_, hasFinally() ? StatementKind::Finally : StatementKind::Try);
-        finallyStart_.offset = 0;
-    }
-
-    // Emits JSOP_GOTO to the end of try-catch-finally.
-    // Used in `yield*`.
-    MOZ_MUST_USE bool emitJumpOverCatchAndFinally() {
-        if (!bce_->emitJump(JSOP_GOTO, &catchAndFinallyJump_))
-            return false;
-        return true;
-    }
-
-    MOZ_MUST_USE bool emitTry() {
-        MOZ_ASSERT(state_ == State::Start);
-
-        // Since an exception can be thrown at any place inside the try block,
-        // we need to restore the stack and the scope chain before we transfer
-        // the control to the exception handler.
-        //
-        // For that we store in a try note associated with the catch or
-        // finally block the stack depth upon the try entry. The interpreter
-        // uses this depth to properly unwind the stack and the scope chain.
-        depth_ = bce_->stackDepth;
-
-        // Record the try location, then emit the try block.
-        if (!bce_->newSrcNote(SRC_TRY, &noteIndex_))
-            return false;
-        if (!bce_->emit1(JSOP_TRY))
-            return false;
-        tryStart_ = bce_->offset();
-
-#ifdef DEBUG
-        state_ = State::Try;
-#endif
-        return true;
-    }
-
-  private:
-    MOZ_MUST_USE bool emitTryEnd() {
-        MOZ_ASSERT(state_ == State::Try);
-        MOZ_ASSERT(depth_ == bce_->stackDepth);
-
-        // GOSUB to finally, if present.
-        if (hasFinally() && controlInfo_) {
-            if (!bce_->emitJump(JSOP_GOSUB, &controlInfo_->gosubs))
-                return false;
-        }
-
-        // Source note points to the jump at the end of the try block.
-        if (!bce_->setSrcNoteOffset(noteIndex_, 0, bce_->offset() - tryStart_ + JSOP_TRY_LENGTH))
-            return false;
-
-        // Emit jump over catch and/or finally.
-        if (!bce_->emitJump(JSOP_GOTO, &catchAndFinallyJump_))
-            return false;
-
-        if (!bce_->emitJumpTarget(&tryEnd_))
-            return false;
-
-        return true;
-    }
-
-  public:
-    MOZ_MUST_USE bool emitCatch() {
-        if (!emittedTry_) {
-            MOZ_ASSERT(state_ == State::Try);
-            if (!emitTryEnd())
-                return false;
-        } else {
-            MOZ_ASSERT(state_ == State::Catch);
-            if (!emitCatchEnd(true))
-                return false;
-        }
-
-        MOZ_ASSERT(bce_->stackDepth == depth_);
-
-        if (controlKind_ == ControlKind::Syntactic) {
-            // Clear the frame's return value that might have been set by the
-            // try block:
-            //
-            //   eval("try { 1; throw 2 } catch(e) {}"); // undefined, not 1
-            if (!bce_->emit1(JSOP_UNDEFINED))
-                return false;
-            if (!bce_->emit1(JSOP_SETRVAL))
-                return false;
-        }
-
-#ifdef DEBUG
-        state_ = State::Catch;
-#endif
-        emittedTry_ = true;
-        return true;
-    }
-
-  private:
-    MOZ_MUST_USE bool emitCatchEnd(bool hasNext) {
-        MOZ_ASSERT(state_ == State::Catch);
-
-        if (!controlInfo_)
-            return true;
-
-        // gosub <finally>, if required.
-        if (hasFinally()) {
-            if (!bce_->emitJump(JSOP_GOSUB, &controlInfo_->gosubs))
-                return false;
-            MOZ_ASSERT(bce_->stackDepth == depth_);
-        }
-
-        // Jump over the remaining catch blocks.  This will get fixed
-        // up to jump to after catch/finally.
-        if (!bce_->emitJump(JSOP_GOTO, &catchAndFinallyJump_))
-            return false;
-
-        // If this catch block had a guard clause, patch the guard jump to
-        // come here.
-        if (controlInfo_->guardJump.offset != -1) {
-            if (!bce_->emitJumpTargetAndPatch(controlInfo_->guardJump))
-                return false;
-            controlInfo_->guardJump.offset = -1;
-
-            // If this catch block is the last one, rethrow, delegating
-            // execution of any finally block to the exception handler.
-            if (!hasNext) {
-                if (!bce_->emit1(JSOP_EXCEPTION))
-                    return false;
-                if (!bce_->emit1(JSOP_THROW))
-                    return false;
-            }
-        }
-
-        return true;
-    }
-
-  public:
-    // If `finallyPos` is specified, it's an offset of the finally block's
-    // "{" character in the source code text, to improve line:column number in
-    // the error reporting.
-    // For non-syntactic try-catch-finally, `finallyPos` can be omitted.
-    MOZ_MUST_USE bool emitFinally(const Maybe<uint32_t>& finallyPos = Nothing()) {
-        // If we are using controlInfo_ (i.e., emitting a syntactic try
-        // blocks), we must have specified up front if there will be a finally
-        // close. For internal non-syntactic try blocks, like those emitted for
-        // yield* and IteratorClose inside for-of loops, we can emitFinally even
-        // without specifying up front, since the internal non-syntactic try
-        // blocks emit no GOSUBs.
-        if (!controlInfo_) {
-            if (kind_ == Kind::TryCatch)
-                kind_ = Kind::TryCatchFinally;
-        } else {
-            MOZ_ASSERT(hasFinally());
-        }
-
-        if (!hasCatch()) {
-            MOZ_ASSERT(state_ == State::Try);
-            if (!emitTryEnd())
-                return false;
-        } else {
-            MOZ_ASSERT(state_ == State::Catch);
-            if (!emitCatchEnd(false))
-                return false;
-        }
-
-        MOZ_ASSERT(bce_->stackDepth == depth_);
-
-        if (!bce_->emitJumpTarget(&finallyStart_))
-            return false;
-
-        if (controlInfo_) {
-            // Fix up the gosubs that might have been emitted before non-local
-            // jumps to the finally code.
-            bce_->patchJumpsToTarget(controlInfo_->gosubs, finallyStart_);
-
-            // Indicate that we're emitting a subroutine body.
-            controlInfo_->setEmittingSubroutine();
-        }
-        if (finallyPos) {
-            if (!bce_->updateSourceCoordNotes(finallyPos.value()))
-                return false;
-        }
-        if (!bce_->emit1(JSOP_FINALLY))
-            return false;
-
-        if (controlKind_ == ControlKind::Syntactic) {
-            if (!bce_->emit1(JSOP_GETRVAL))
-                return false;
-
-            // Clear the frame's return value to make break/continue return
-            // correct value even if there's no other statement before them:
-            //
-            //   eval("x: try { 1 } finally { break x; }"); // undefined, not 1
-            if (!bce_->emit1(JSOP_UNDEFINED))
-                return false;
-            if (!bce_->emit1(JSOP_SETRVAL))
-                return false;
-        }
-
-#ifdef DEBUG
-        state_ = State::Finally;
-#endif
-        return true;
-    }
-
-  private:
-    MOZ_MUST_USE bool emitFinallyEnd() {
-        MOZ_ASSERT(state_ == State::Finally);
-
-        if (controlKind_ == ControlKind::Syntactic) {
-            if (!bce_->emit1(JSOP_SETRVAL))
-                return false;
-        }
-
-        if (!bce_->emit1(JSOP_RETSUB))
-            return false;
-
-        bce_->hasTryFinally = true;
-        return true;
-    }
-
-  public:
-    MOZ_MUST_USE bool emitEnd() {
-        if (!hasFinally()) {
-            MOZ_ASSERT(state_ == State::Catch);
-            if (!emitCatchEnd(false))
-                return false;
-        } else {
-            MOZ_ASSERT(state_ == State::Finally);
-            if (!emitFinallyEnd())
-                return false;
-        }
-
-        MOZ_ASSERT(bce_->stackDepth == depth_);
-
-        // ReconstructPCStack needs a NOP here to mark the end of the last
-        // catch block.
-        if (!bce_->emit1(JSOP_NOP))
-            return false;
-
-        // Fix up the end-of-try/catch jumps to come here.
-        if (!bce_->emitJumpTargetAndPatch(catchAndFinallyJump_))
-            return false;
-
-        // Add the try note last, to let post-order give us the right ordering
-        // (first to last for a given nesting level, inner to outer by level).
-        if (hasCatch()) {
-            if (!bce_->tryNoteList.append(JSTRY_CATCH, depth_, tryStart_, tryEnd_.offset))
-                return false;
-        }
-
-        // If we've got a finally, mark try+catch region with additional
-        // trynote to catch exceptions (re)thrown from a catch block or
-        // for the try{}finally{} case.
-        if (hasFinally()) {
-            if (!bce_->tryNoteList.append(JSTRY_FINALLY, depth_, tryStart_, finallyStart_.offset))
-                return false;
-        }
-
-#ifdef DEBUG
-        state_ = State::End;
-#endif
-        return true;
-    }
-};
 
 // OOOEEE start
 
@@ -662,204 +207,6 @@ class MOZ_RAII BytecodeEmitter::OptionalEmitter {
 };
 
 // OOOEEE end
-
-class ForOfLoopControl : public LoopControl
-{
-    // The stack depth of the iterator.
-    int32_t iterDepth_;
-
-    // for-of loops, when throwing from non-iterator code (i.e. from the body
-    // or from evaluating the LHS of the loop condition), need to call
-    // IteratorClose.  This is done by enclosing non-iterator code with
-    // try-catch and call IteratorClose in `catch` block.
-    // If IteratorClose itself throws, we must not re-call IteratorClose. Since
-    // non-local jumps like break and return call IteratorClose, whenever a
-    // non-local jump is emitted, we must tell catch block not to perform
-    // IteratorClose.
-    //
-    //   for (x of y) {
-    //     // Operations for iterator (IteratorNext etc) are outside of
-    //     // try-block.
-    //     try {
-    //       ...
-    //       if (...) {
-    //         // Before non-local jump, clear iterator on the stack to tell
-    //         // catch block not to perform IteratorClose.
-    //         tmpIterator = iterator;
-    //         iterator = undefined;
-    //         IteratorClose(tmpIterator, { break });
-    //         break;
-    //       }
-    //       ...
-    //     } catch (e) {
-    //       // Just throw again when iterator is cleared by non-local jump.
-    //       if (iterator === undefined)
-    //         throw e;
-    //       IteratorClose(iterator, { throw, e });
-    //     }
-    //   }
-    Maybe<TryEmitter> tryCatch_;
-
-    // Used to track if any yields were emitted between calls to to
-    // emitBeginCodeNeedingIteratorClose and emitEndCodeNeedingIteratorClose.
-    uint32_t numYieldsAtBeginCodeNeedingIterClose_;
-
-    bool allowSelfHosted_;
-
-    IteratorKind iterKind_;
-
-  public:
-    ForOfLoopControl(BytecodeEmitter* bce, int32_t iterDepth, bool allowSelfHosted,
-                     IteratorKind iterKind)
-      : LoopControl(bce, StatementKind::ForOfLoop),
-        iterDepth_(iterDepth),
-        numYieldsAtBeginCodeNeedingIterClose_(UINT32_MAX),
-        allowSelfHosted_(allowSelfHosted),
-        iterKind_(iterKind)
-    {
-    }
-
-    bool emitBeginCodeNeedingIteratorClose(BytecodeEmitter* bce) {
-        tryCatch_.emplace(bce, TryEmitter::Kind::TryCatch, TryEmitter::ControlKind::NonSyntactic);
-
-        if (!tryCatch_->emitTry())
-            return false;
-
-        MOZ_ASSERT(numYieldsAtBeginCodeNeedingIterClose_ == UINT32_MAX);
-        numYieldsAtBeginCodeNeedingIterClose_ = bce->yieldAndAwaitOffsetList.numYields;
-
-        return true;
-    }
-
-    bool emitEndCodeNeedingIteratorClose(BytecodeEmitter* bce) {
-        if (!tryCatch_->emitCatch())              // ITER ...
-            return false;
-
-        if (!bce->emit1(JSOP_EXCEPTION))          // ITER ... EXCEPTION
-            return false;
-        unsigned slotFromTop = bce->stackDepth - iterDepth_;
-        if (!bce->emitDupAt(slotFromTop))         // ITER ... EXCEPTION ITER
-            return false;
-
-        // If ITER is undefined, it means the exception is thrown by
-        // IteratorClose for non-local jump, and we should't perform
-        // IteratorClose again here.
-        if (!bce->emit1(JSOP_UNDEFINED))          // ITER ... EXCEPTION ITER UNDEF
-            return false;
-        if (!bce->emit1(JSOP_STRICTNE))           // ITER ... EXCEPTION NE
-            return false;
-
-        InternalIfEmitter ifIteratorIsNotClosed(bce);
-        if (!ifIteratorIsNotClosed.emitThen())    // ITER ... EXCEPTION
-            return false;
-
-        MOZ_ASSERT(slotFromTop == unsigned(bce->stackDepth - iterDepth_));
-        if (!bce->emitDupAt(slotFromTop))         // ITER ... EXCEPTION ITER
-            return false;
-        if (!emitIteratorCloseInInnermostScope(bce, CompletionKind::Throw))
-            return false;                         // ITER ... EXCEPTION
-
-        if (!ifIteratorIsNotClosed.emitEnd())     // ITER ... EXCEPTION
-            return false;
-
-        if (!bce->emit1(JSOP_THROW))              // ITER ...
-            return false;
-
-        // If any yields were emitted, then this for-of loop is inside a star
-        // generator and must handle the case of Generator.return. Like in
-        // yield*, it is handled with a finally block.
-        uint32_t numYieldsEmitted = bce->yieldAndAwaitOffsetList.numYields;
-        if (numYieldsEmitted > numYieldsAtBeginCodeNeedingIterClose_) {
-            if (!tryCatch_->emitFinally())
-                return false;
-
-            InternalIfEmitter ifGeneratorClosing(bce);
-            if (!bce->emit1(JSOP_ISGENCLOSING))   // ITER ... FTYPE FVALUE CLOSING
-                return false;
-            if (!ifGeneratorClosing.emitThen())   // ITER ... FTYPE FVALUE
-                return false;
-            if (!bce->emitDupAt(slotFromTop + 1)) // ITER ... FTYPE FVALUE ITER
-                return false;
-            if (!emitIteratorCloseInInnermostScope(bce, CompletionKind::Normal))
-                return false;
-            if (!ifGeneratorClosing.emitEnd())    // ITER ... FTYPE FVALUE
-                return false;
-        }
-
-        if (!tryCatch_->emitEnd())
-            return false;
-
-        tryCatch_.reset();
-        numYieldsAtBeginCodeNeedingIterClose_ = UINT32_MAX;
-
-        return true;
-    }
-
-    bool emitIteratorCloseInInnermostScope(BytecodeEmitter* bce,
-                                           CompletionKind completionKind = CompletionKind::Normal) {
-        return emitIteratorCloseInScope(bce,  *bce->innermostEmitterScope(), completionKind);
-    }
-
-    bool emitIteratorCloseInScope(BytecodeEmitter* bce,
-                                  EmitterScope& currentScope,
-                                  CompletionKind completionKind = CompletionKind::Normal) {
-        ptrdiff_t start = bce->offset();
-        if (!bce->emitIteratorCloseInScope(currentScope, iterKind_, completionKind,
-                                           allowSelfHosted_))
-        {
-            return false;
-        }
-        ptrdiff_t end = bce->offset();
-        return bce->tryNoteList.append(JSTRY_FOR_OF_ITERCLOSE, 0, start, end);
-    }
-
-    bool emitPrepareForNonLocalJumpFromScope(BytecodeEmitter* bce,
-                                             EmitterScope& currentScope,
-                                             bool isTarget) {
-        // Pop unnecessary value from the stack.  Effectively this means
-        // leaving try-catch block.  However, the performing IteratorClose can
-        // reach the depth for try-catch, and effectively re-enter the
-        // try-catch block.
-        if (!bce->emit1(JSOP_POP))                        // ITER
-            return false;
-
-        // Clear ITER slot on the stack to tell catch block to avoid performing
-        // IteratorClose again.
-        if (!bce->emit1(JSOP_UNDEFINED))                  // ITER UNDEF
-            return false;
-        if (!bce->emit1(JSOP_SWAP))                       // UNDEF ITER
-            return false;
-
-        if (!emitIteratorCloseInScope(bce, currentScope, CompletionKind::Normal)) // UNDEF
-            return false;
-
-        if (isTarget) {
-            // At the level of the target block, there's bytecode after the
-            // loop that will pop the iterator and the value, so push
-            // an undefined to balance the stack.
-            if (!bce->emit1(JSOP_UNDEFINED))              // UNDEF UNDEF
-                return false;
-        } else {
-            if (!bce->emit1(JSOP_POP))                    //
-                return false;
-        }
-
-        return true;
-    }
-};
-
-namespace js {
-namespace frontend {
-
-template <>
-bool
-NestableControl::is<ForOfLoopControl>() const
-{
-    return kind_ == StatementKind::ForOfLoop;
-}
-
-} // namespace frontend
-} // namespace js
 
 BytecodeEmitter::BytecodeEmitter(BytecodeEmitter* parent,
                                  const EitherParser<FullParseHandler>& parser, SharedContext* sc,
